refactor: move plugins to plugins/cmd
