+++ plugins/cmd/depl_calls_svc/Dockerfile
