+++ plugins/cmd/litellm/Dockerfile
