+++ plugins/cmd/litellm/plugin.go
